git/git.go: make RefBeforeFirstCommit a *git.Ref
To make a further refactor simpler, let's change the type of
git.RefBeforeFirstCommit from a string (containing a hexadecimal-encoded
version of the SHA-1 reference before a first commit) to a *git.Ref of
the same effect.

Since the reference is not explicitly named, we'll avoid naming it in
the initialization.

An unfortunate side-effect is that declaring a heap-allocated git.Ref
causes us to move the definition from a const- to var-block, which means
that callers can re-assign it at runtime.

Between this, and having to constantly address the value with &, the
former is preferable.