commands/command_ls_files.go: initialize currentRef early
In a subsequent commit, we will compare the current reference (or, in
the case of repositories without a HEAD, the pseudo-reference before the
first commit) to the given reference.

In order to do so in all cases, rather than just the one where
"len(args) != 1", let's force the initialization of this variable
earlier, such that it is available when we need it.