commands/command_ls_files: hide index with arguments
Since e2717688 (commands: scan repository index in git-lfs-ls-files(1),
2017-12-19), we have traversed the contents of the index when showing
files in 'git lfs ls-files'.

The purpose of this was to match behavior in 'git ls-files', where the
index is also shown by default.

However, when given an argument, 'git lfs ls-files' will traverse
starting from a specific revision, so showing the index unconditionally
causes a problem here. It does not make sense to show the current state
of the index when listing files from HEAD~N (where N is non-zero).

So, let's keep track of both the current revision, and the revision that
we were asked to show, and only scan the index when the two are the
same.